Vegetable eat. Duck Caterpillar

+++ src/com/org/model/interfaces/Herbivores.java
@@ -4,5 +4,5 @@ package com.org.model.interfaces;
 import com.org.model.island.vegetable.Vegatable;
 
 public interface Herbivores {
-    Double eatVeggie(Vegatable vegatable);
+    Float eatVeggie(Vegatable vegatable);
 }
